// src/app/contracts/models/diagram-instance.model.ts
export class DiagramInstance {

    id: string;
    parentId: string;
    isVisible: boolean;

    diagram: any;
    moddle: any;
    canvas: any;
    elementFactory: any;
    graphicsFactory: any;
    eventBus: any;
    root: any;
    modelerRenderer: any;
    selection: any;
    elementRegistry: any;
    styles: any;
    keyboard: any;
    minimap: any;
    copyPaste: any;
    modeling: any;
    connect: any;
    popupMenu: any;
    gridSnapping: any;
    snapping: any;
    desktopData: any;
    dragging: any;
    textElements: any;
    translate: any;
    alignElements: any;
    distributeElements: any;
    clipboard: any;
    editorActions: any;
    diagramRegistry: any;
    elementIcons: any;
    viewEventsHandler: any;
}